Modularize event blueprints
